requirements.txt
numpy==1.21.6
pandas==1.3.5
Pillow==9.0.1
mindspore-gpu==1.6.0
python==3.6.0